Fix float32 dtype issue in magnitude calculations (issue #953)
When star particle ages and metals were stored as float32, the magnitude
calculation could return NaN values due to a dtype mismatch in the clamping
mechanism within SSPTable.interpolate(). This occurred because the SSP
tables use float64 internally, but the clamping with np.nextafter() was
failing when comparing float32 values against float64 table bounds.

The fix converts ages and metals arrays to float64 before passing them to
the interpolation routine, ensuring consistent dtype handling throughout
the calculation.

Also adds a regression test that verifies magnitude calculations work
correctly with both float32 and float64 input arrays, and that no NaN
values are produced.

Fixes #953

+++ pynbody/analysis/luminosity.py
@@ -201,9 +201,9 @@ class SSPTable:
 
         """
 
-        age_star = snapshot['age'].in_units('yr')
+        age_star = snapshot['age'].in_units('yr').astype(np.float64)
         age_star[age_star<1.0] = 1.0
-        metals = snapshot['metals']
+        metals = snapshot['metals'].astype(np.float64)
         try:
             masses = snapshot['massform'].in_units('Msol')
         except KeyError:
